titanic-data-project/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
python-dotenv
pyarrow
